backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas==2.1.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic-settings==2.1.0
pydantic==2.5.0
pydantic[email]==2.5.0
httpx==0.24.1
aiohttp==3.9.4
bcrypt==4.0.1
passlib==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0
sqlalchemy==2.0.23